requirements.txt
certifi
fastapi
pydantic
pytest
python-dotenv
requests
requests_toolbelt
typing
uvicorn
httpx
pytz
numpy
fastapi_utils